Adicionados os testes de unidade e de integração

+++ src/main/java/com/waynesouza/soccer/domain/dto/PartidaAtualizadaDTO.java
@@ -2,6 +2,7 @@ package com.waynesouza.soccer.domain.dto;
 
 import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.Min;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -23,7 +24,7 @@ public class PartidaAtualizadaDTO implements Serializable {
 
     @Min(value = 0, message = "O valor deve ser um número inteiro maior que zero")
     @Digits(integer = 2, fraction = 0, message = "O valor deve ser um número inteiro positivo ou zero")
-    private Integer quantidadeGolVisitante;
+    private BigDecimal quantidadeGolVisitante;
 
     private LocalDateTime dataHora;
