+++ src/main/java/com/waynesouza/soccer/domain/dto/PartidaDTO.java
@@ -3,10 +3,12 @@ package com.waynesouza.soccer.domain.dto;
 import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -21,7 +23,7 @@ public class PartidaDTO implements Serializable {
     @NotNull(message = "A quantidade de gols do time mandante deve ser informada")
     @Min(value = 0, message = "O valor deve ser um número inteiro maior que zero")
     @Digits(integer = 2, fraction = 0, message = "O valor deve ser um número inteiro positivo ou zero")
-    private Integer quantidadeGolMandante;
+    private BigDecimal quantidadeGolMandante;
 
     @NotNull(message = "O time visitante deve ser informado")
     private String timeVisitante;
@@ -29,7 +31,7 @@ public class PartidaDTO implements Serializable {
     @NotNull(message = "A quantidade de gols do time visitante deve ser informada")
     @Min(value = 0, message = "O valor deve ser um número inteiro maior que zero")
     @Digits(integer = 2, fraction = 0, message = "O valor deve ser um número inteiro positivo ou zero")
-    private Integer quantidadeGolVisitante;
+    private BigDecimal quantidadeGolVisitante;
 
     @NotNull(message = "A data com o horário da partida deve ser informada")
     private LocalDateTime dataHora;
